Check if file/dir exists before unlink()ing
git-svn-id: http://svn.php.net/repository/pear/packages/HTTP_Session2/trunk@295133 c90b9560-bf6c-de11-be94-00142212c4b1

+++ tests/test-session-doctrine.phpt
@@ -77,8 +77,12 @@ try {
 <?php
 $_tmp = dirname(__FILE__) . '/tmp';
 $_db  = $_tmp . '/test.db';
-unlink($_db);
-rmdir($_tmp);
+if (file_exists($_db)) {
+    unlink($_db);
+}
+if (dir_exists($_tmp)) {
+    rmdir($_tmp);
+}
 --EXPECT--
 string(9) "Setting.."
 string(12) "Retrieving.."
